+++ frontend/src/components/Lastproductshome/slider/Leftcontrol.js
