+++ frontend/src/components/Lastproductshome/slider/Rightcontrol.js
