+++ frontend/src/components/Lastproductshome/slider/Slide.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+function Slide({ children }) {
+	return <div className='slide'>{children}</div>
+}
+
+export default Slide
